feat: Refactor error handling for image operations and improve exception naming
