+++ src/ai/backend/manager/repositories/image/admin_repository.py
@@ -1,4 +1,3 @@
-from typing import Optional
 from uuid import UUID
 
 from sqlalchemy.exc import DBAPIError
@@ -8,6 +7,7 @@ from ai.backend.common.docker import ImageRef
 from ai.backend.common.types import ImageAlias
 from ai.backend.manager.data.image.types import ImageData
 from ai.backend.manager.errors.exceptions import (
+    ImageNotFound,
     PurgeImageActionByIdObjectDBError,
 )
 from ai.backend.manager.models.image import (
@@ -37,8 +37,11 @@ class AdminImageRepository:
 
     async def _get_image_by_id(
         self, session: SASession, image_id: UUID, load_aliases: bool = False
-    ) -> Optional[ImageRow]:
-        return await ImageRow.get(session, image_id, load_aliases=load_aliases)
+    ) -> ImageRow:
+        image_row = await ImageRow.get(session, image_id, load_aliases=load_aliases)
+        if not image_row:
+            raise ImageNotFound()
+        return image_row
 
     async def soft_delete_image_force(
         self, identifiers: list[ImageAlias | ImageRef | ImageIdentifier]
@@ -60,13 +63,11 @@ class AdminImageRepository:
         """
         async with self._db.begin_session() as session:
             image_row = await self._get_image_by_id(session, image_id)
-            if not image_row:
-                raise ValueError(f"Image with ID {image_id} not found")
             await image_row.mark_as_deleted(session)
             data = image_row.to_dataclass()
         return data
 
-    async def delete_image_with_aliases_force(self, image_id: UUID) -> Optional[ImageData]:
+    async def delete_image_with_aliases_force(self, image_id: UUID) -> ImageData:
         """
         Deletes an image and all its aliases without checking ownership.
         This is a forceful deletion and should be used with caution.
@@ -74,8 +75,6 @@ class AdminImageRepository:
         try:
             async with self._db.begin_session() as session:
                 image_row = await self._get_image_by_id(session, image_id, load_aliases=True)
-                if not image_row:
-                    return None
                 data = image_row.to_dataclass()
                 for alias in image_row.aliases:
                     await session.delete(alias)
@@ -84,15 +83,13 @@ class AdminImageRepository:
         except DBAPIError as e:
             raise PurgeImageActionByIdObjectDBError(str(e))
 
-    async def untag_image_from_registry_force(self, image_id: UUID) -> Optional[ImageData]:
+    async def untag_image_from_registry_force(self, image_id: UUID) -> ImageData:
         """
         Untags an image from registry without ownership check.
         This is an admin-only operation.
         """
         async with self._db.begin_readonly_session() as session:
             image_row = await self._get_image_by_id(session, image_id, load_aliases=True)
-            if not image_row:
-                return None
             await image_row.untag_image_from_registry(self._db, session)
             data = image_row.to_dataclass()
         return data
