+++ src/ai/backend/manager/services/image/actions/forget_image_by_id.py
@@ -51,16 +51,3 @@ class ForgetImageActionByIdGenericForbiddenError(BackendAIError):
             operation=ErrorOperation.SOFT_DELETE,
             error_detail=ErrorDetail.FORBIDDEN,
         )
-
-
-class ForgetImageActionByIdObjectNotFoundError(BackendAIError):
-    error_type = "https://api.backend.ai/probs/image-not-found"
-    error_title = "Image not found."
-
-    @classmethod
-    def error_code(cls) -> ErrorCode:
-        return ErrorCode(
-            domain=ErrorDomain.IMAGE,
-            operation=ErrorOperation.SOFT_DELETE,
-            error_detail=ErrorDetail.NOT_FOUND,
-        )
