+++ tests/manager/services/image/actions/test_modify_image.py
@@ -4,13 +4,12 @@ import pytest
 
 from ai.backend.common.types import SlotName
 from ai.backend.manager.data.image.types import ImageLabelsData, ImageResourcesData
-from ai.backend.manager.errors.exceptions import ImageNotFound
+from ai.backend.manager.errors.exceptions import ImageNotFound, ModifyImageActionValueError
 from ai.backend.manager.models.image import ImageType
 from ai.backend.manager.services.image.actions.modify_image import (
     ImageModifier,
     ModifyImageAction,
     ModifyImageActionResult,
-    ModifyImageActionValueError,
 )
 from ai.backend.manager.services.image.processors import ImageProcessors
 from ai.backend.manager.types import OptionalState, TriState
