+++ tests/manager/services/image/actions/test_purge_image_by_id.py
@@ -3,12 +3,11 @@ import uuid
 import pytest
 import sqlalchemy as sa
 
+from ai.backend.manager.errors.exceptions import ForgetImageForbiddenError, ImageNotFound
 from ai.backend.manager.models.image import ImageRow
 from ai.backend.manager.models.user import UserRole
 from ai.backend.manager.models.utils import ExtendedAsyncSAEngine
 from ai.backend.manager.services.image.actions.purge_image_by_id import (
-    PurgeImageActionByIdGenericForbiddenError,
-    PurgeImageActionByIdObjectNotFoundError,
     PurgeImageByIdAction,
     PurgeImageByIdActionResult,
 )
@@ -41,7 +40,7 @@ from ...test_utils import TestScenario
                 client_role=UserRole.USER,
                 image_id=IMAGE_ROW_FIXTURE.id,
             ),
-            PurgeImageActionByIdGenericForbiddenError,
+            ForgetImageForbiddenError,
         ),
         TestScenario.failure(
             "Image not found",
@@ -50,7 +49,7 @@ from ...test_utils import TestScenario
                 client_role=UserRole.SUPERADMIN,
                 image_id=uuid.UUID("00000000-0000-0000-0000-000000000000"),  # wrong image_id
             ),
-            PurgeImageActionByIdObjectNotFoundError,
+            ImageNotFound,
         ),
     ],
 )
